[Objects Version] Tweak bulk loading logic
## Description

Changes the parametrisation of the `objects_version` bulk loader.
Previously, it used 20 workers, and processed batches of 1M
checkpoints at a time, per partition, now it processes 100K
checkpoints at a time, per partition, using 50 workers.

This change was made after recognising that for hot periods on the
network, the bulk loader was wasting work by timing out multiple times
on sections until it eventually settled on a batch size of around 100K.

## Test plan

Run a bulk load and see that it takes less time than before.