[Bulk Load] 1-GIN Benchmark (#14)
## Description

Simulate paginated transaction queries over the new proposed 1-GIN
schema, and run queries over them.

## Test plan

Ran the benchmarks. For now the results seem to be comparable or
slightly better than what the existing schema supports. Below are the
success rates for queries involving each kind of filter, using the
1-GIN schema:

```
([#{:mod} 0.6086956521739131]
 [#{:fun} 0.64]
 [#{:pkg} 0.6419753086419753]
 [#{:changed} 0.7058823529411765]
 [#{:kind} 0.7717391304347826]
 [#{:cp-<} 0.7857142857142857]
 [#{:input} 0.8]
 [#{:before} 0.8170731707317073]
 [#{:ids} 1.0]
 [#{:cp->} 1.0]
 [#{:after} 1.0]
 [#{:cp-=} 1.0])
```

And the same for the existing schema:

```
([#{:mod} 0.6086956521739131]
 [#{:pkg} 0.6190476190476191]
 [#{:fun} 0.64]
 [#{:changed} 0.6486486486486487]
 [#{:kind} 0.7553191489361702]
 [#{:cp-<} 0.7692307692307693]
 [#{:input} 0.8]
 [#{:before} 0.8023952095808383]
 [#{:ids} 1.0]
 [#{:cp->} 1.0]
 [#{:after} 1.0]
 [#{:cp-=} 1.0])
```

It's possible that the dominating factor here is that queries that
need to scan the whole history are not feasible rather than the issue
being the number of filters that are being combined.