[Footprint] Simulate Abstracted IDs
## Description

Simulate replacing 32B IDs with an 8B key.

The number of unique addresses and IDs is estimated by summing the
number of tuples in `objects_snapshot` (which includes deleted IDs),
and the number of distinct addresses in `tx_recipients`. This in turn
is used to estimate the size of a new lookaside table mapping long IDs
to short IDs.

The `id-columns` data structure tracks the write amplification of
columns that include Object IDs or Addresses. Every such instance of a
column nets a 24B size saving.

## Test plan

```
(use 'footprint)
(def db ...)
(def stats (table-stats db))

(->> stats abstract-ids footprint pretty-size)
```