[Footprint] Simulate a single event sequence number
## Description

What if we could represent event sequence numbers using a single
number, rather than a pair of transaction and event sequence number?

Also introduced a single entrypoint for turning optimizations on and
off. This was to deal with the complexity that ordering of
optimizations matters (some optimizations introduce statistics that
other optimizations then need to deal with.)

Also hardened the detection of an empty `tuples` column -- this column
can contain a `-1` if postgres has not yet analyzed it.

## Test plan

```
(use 'footprint)
(def stats (table-stats db))

(-> stats (optimize :event-seq) footprint pretty-size)
```