[Footprint] Estimate effect of offloading to KV store (#29)
## Description

Simulate offloading fields to the key-value store. The simulation
selects fields that don't participate in filtering (they are read, but
not filtered against in SELECTs) as candidates for offloading and
estimates the size of these columns using the `pg_stats` table (average
width) combined with the `pg_class` table (number of tuples).

Additionally, we assume that:

- If a table has candidate columns for the KV-store, then its TOAST
table is also assumed to have been offloaded. The TOAST table contains
the contents of values that are too big to store directly on the heap.
It is not broken down by column, but there's a good chance that if a
table has columns that are suitable for offloading to a KV-store, they
will be the ones that contain TOASTed values. (Also note that the
average width statistic for columns does not account for the size of
TOASTed values, just the forwarding pointer in the heap that points to
the TOASTed value).
- If all of a table's columns get moved to the KV-store, then we can get
rid of it entirely (and all its indices).

## Test plan

```
(def db ...)

(use 'footprint)
(def stats (table-stats db))
(->> stats offloaded-to-kv-store footprint pretty-size)
```

## Stack

- #21 
- #22 
- #23 
- #24 
- #25 
- #26 
- #27 
- #28